+++ lib/base/streams.h
@@ -8,7 +8,7 @@
 template <typename T>
 std::ostream& operator<<(std::ostream& os, const std::vector<T>& vec) {
   os << "[";
-  for (int i = 0; i < vec.size(); ++i) {
+  for (unsigned long i = 0; i < vec.size(); ++i) {
     if (i != 0) {
       os << ",";
     }
